Se cambia el tipo de dato

+++ Sofka.Piguinera.Cotizacion/Sofka.Piguinera.Cotizacion/Models/Strategy/IDiscountStrategy.cs
@@ -4,7 +4,7 @@
     {
 
         bool CanApply(int seniority);
-        decimal Apply();
+        float Apply();
 
     }
 }
